+++ packages/types/src/aztec/index.ts
@@ -1,4 +1,5 @@
 export * from "./l2Block.js";
 export * from "./l2Contract.js";
+export * from "./l2TxEffect.js";
 
 export { type NodeInfo } from "@aztec/aztec.js";
